Move item-specific stuff from suggestions-generation
