+++ src/app/search/suggestions-generation.ts
@@ -1,61 +1,4 @@
-import { CustomStatDef } from '@destinyitemmanager/dim-api-types';
-import { D2ManifestDefinitions } from 'app/destiny2/d2-definitions';
-import { customStatsSelector } from 'app/dim-api/selectors';
-import { TagValue } from 'app/inventory/dim-item-info';
-import { DimItem } from 'app/inventory/item-types';
-import { Loadout } from 'app/loadout-drawer/loadout-types';
-import { loadoutsSelector } from 'app/loadout-drawer/loadouts-selector';
-import { d2ManifestSelector } from 'app/manifest/selectors';
-import { createSelector } from 'reselect';
-import {
-  allItemsSelector,
-  allNotesHashtagsSelector,
-  getNotesSelector,
-  getTagSelector,
-} from '../inventory/selectors';
-
 import { FilterDefinition, canonicalFilterFormats } from './filter-types';
-import { SuggestionsContext } from './item-filter-types';
-
-//
-// Selectors
-//
-
-/**
- * A selector for the suggestionsContext for a particular destiny version.
- * This must depend on every bit of data in suggestionsContext so that we
- * regenerate filter suggestions whenever any of them changes.
- */
-export const suggestionsContextSelector = createSelector(
-  allItemsSelector,
-  loadoutsSelector,
-  d2ManifestSelector,
-  getTagSelector,
-  getNotesSelector,
-  allNotesHashtagsSelector,
-  customStatsSelector,
-  makeSuggestionsContext,
-);
-
-function makeSuggestionsContext(
-  allItems: DimItem[],
-  loadouts: Loadout[],
-  d2Definitions: D2ManifestDefinitions | undefined,
-  getTag: (item: DimItem) => TagValue | undefined,
-  getNotes: (item: DimItem) => string | undefined,
-  allNotesHashtags: string[],
-  customStats: CustomStatDef[],
-): SuggestionsContext {
-  return {
-    allItems,
-    loadouts,
-    d2Definitions,
-    getTag,
-    getNotes,
-    allNotesHashtags,
-    customStats,
-  };
-}
 
 const operators = ['<', '>', '<=', '>=']; // TODO: add "none"? remove >=, <=?
